requirements.txt
python-telegram-bot
yfinance
requests
flask